chore: Eliminate code duplication
