+++ packages/next/src/client/components/not-found.ts
@@ -1,6 +1,6 @@
-const NOT_FOUND_ERROR_CODE = 'NEXT_NOT_FOUND'
+import { createUIError } from './ui-error-builder'
 
-type NotFoundError = Error & { digest: typeof NOT_FOUND_ERROR_CODE }
+const { thrower, matcher } = createUIError('NEXT_NOT_FOUND')
 
 /**
  * This function allows you to render the [not-found.js file](https://nextjs.org/docs/app/api-reference/file-conventions/not-found)
@@ -16,12 +16,7 @@ type NotFoundError = Error & { digest: typeof NOT_FOUND_ERROR_CODE }
  *
  * Read more: [Next.js Docs: `notFound`](https://nextjs.org/docs/app/api-reference/functions/not-found)
  */
-export function notFound(): never {
-  // eslint-disable-next-line no-throw-literal
-  const error = new Error(NOT_FOUND_ERROR_CODE)
-  ;(error as NotFoundError).digest = NOT_FOUND_ERROR_CODE
-  throw error
-}
+const notFound = thrower
 
 /**
  * Checks an error to determine if it's an error generated by the `notFound()`
@@ -30,10 +25,6 @@ export function notFound(): never {
  * @param error the error that may reference a not found error
  * @returns true if the error is a not found error
  */
-export function isNotFoundError(error: unknown): error is NotFoundError {
-  if (typeof error !== 'object' || error === null || !('digest' in error)) {
-    return false
-  }
+const isNotFoundError = matcher
 
-  return error.digest === NOT_FOUND_ERROR_CODE
-}
+export { isNotFoundError, notFound }
