+++ packages/next/src/server/app-render/entry-base.ts
@@ -18,8 +18,6 @@ import {
   createDynamicallyTrackedSearchParams,
 } from '../../client/components/search-params'
 import * as serverHooks from '../../client/components/hooks-server-context'
-import { NotFoundBoundary } from '../../client/components/not-found-boundary'
-import { ForbiddenBoundary } from '../../client/components/forbidden-boundary'
 import { patchFetch as _patchFetch } from '../lib/patch-fetch'
 // not being used but needs to be included in the client manifest for /_not-found
 import '../../client/components/error-boundary'
@@ -31,6 +29,10 @@ import {
 } from '../../server/app-render/rsc/preloads'
 import { Postpone } from '../../server/app-render/rsc/postpone'
 import { taintObjectReference } from '../../server/app-render/rsc/taint'
+import {
+  ForbiddenBoundary,
+  NotFoundBoundary,
+} from '../../client/components/ui-errors-boundaries'
 
 // patchFetch makes use of APIs such as `React.unstable_postpone` which are only available
 // in the experimental channel of React, so export it from here so that it comes from the bundled runtime
